Replace Celery worker by systemd timer
