+++ src/scripts/hades
@@ -37,7 +37,11 @@ print_usage() {
 }
 
 run_agent() {
-    exec python3 -m celery.bin.worker --beat --app=hades.agent --uid="${HADES_AGENT_USER}" --gid="${HADES_AGENT_GROUP}" --workdir="${HADES_AGENT_HOME}"
+    if [[ $(id -u) = 0 ]]; then
+        exec python3 -m hades.common.su "${HADES_AGENT_USER}" "$0" agent "$@"
+    fi
+    cd "${HADES_AGENT_HOME}"
+    exec python3 -m hades.agent
 }
 
 run_auth_dns() {
