+++ systemd/hades.target
@@ -1,7 +1,7 @@
 [Unit]
 Description=Hades
 Documentation=https://agdsn.github.io/hades/
-Requires=hades-agent.service
+Requires=hades-agent.timer
 Requires=hades-auth-dhcp.service
 Requires=hades-auth-dns.service
 Requires=hades-auth-vrrp.service
